Move lstm data directory and split lstm datagen from mod.rs.
