+++ provider/datagen/data/lstm/lstm_th.json
