+++ provider/datagen/src/error.rs
@@ -29,7 +29,3 @@ pub fn is_missing_icuexport_error(mut e: DataError) -> bool {
 pub(crate) fn data_error_from_toml(other: toml::de::Error) -> DataError {
     DataError::custom("Toml deserialize").with_display_context(&other)
 }
-
-pub(crate) fn data_error_from_json(other: serde_json::Error) -> DataError {
-    DataError::custom("JSON deserialize").with_display_context(&other)
-}
